Create controllers and Cancel

+++ use-cases/use-case-port/src/main/kotlin/com/br/gook/data/input/LocalInputPort.kt
@@ -2,5 +2,5 @@ package com.br.gook.data.input
 
 data class LocalInputPort(
     val name: String,
-    val addressId: Int
+    val addressId: Long
 )
